requirements.txt
amqp==5.3.1
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
black==24.10.0
celery==5.4.0
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.10.3
cryptography==45.0.6
daphne==4.2.1
decorator==5.2.1
diff-match-patch==20241021
Django==5.1.7
django-cors-headers==4.5.0
django-extensions==3.2.3
django-filter==24.3
django-import-export==4.3.9
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dlib==19.24.6
drf-yasg==1.21.9
et_xmlfile==2.0.0
executing==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
flake8==7.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
ipython==8.30.0
isort==5.13.2
jedi==0.19.2
kombu==5.5.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy_extensions==1.1.0
ndg-httpsclient==0.5.1
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pusher==3.3.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.22
pydot==4.0.1
pydotplus==2.0.2
pyflakes==3.2.0
Pygments==2.19.2
pygraphviz==1.14
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.4
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tablib==3.8.0
traitlets==5.14.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
zope.interface==7.2
